requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
openai==1.3.5
python-dateutil==2.8.2
speedtest-cli==2.1.3
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0

